feat: Make most system packages user packages
Making a package a system package should be the exception rather than
the rule. A package should only become a system package if it makes
sense for it to be available to every user of the system. If I imagine
that I would give some other person access to log into my system without
wheel access, I would not want to litter that user's executable
namespace with tools of my own preference, e.g. Neovim or Chromium.