refactor: Rename display/ -> bin/
I changed my mind. Instead of having a bunch of scripts scattered all
over the repo and having to symlink each one in ~/bin/, let's just
collect all such script in a single directory and symlink the ~/bin/
directory itself. Then I can just drop a script in ~/nixos-config/bin/
and it will be available in the PATH, with the added benefit of it being
tracked in source control.