fix(bin): Fix keyboard layout toggling script
Darn, I just realized that all this work setting up this "sesebel"
layout has been for pretty much nothing. I never had to create a
slightly altered variant of the "se" layout after all. All that matters
is that the "section" key is the same on both solemak and the
alternative layout.

But eh, let's keep it as it is. It might be good for reproducibility's
sake to make a "pinned" copy anyway. Or maybe I would want to change the
sesebel layout in the future, I dunno.