fix(keyboard/solaar): Try fixing sporadic chromium close issue
About 1 % of the time, when trying to close a Chromium tab using the
Forward button on my mouse, what happens instead is that the whole
window closes. My hypothesis is that the Solaar rule that closes windows
gets triggered instead of the rule that closes tabs. Explicitly stating
that the window must never be closed if the process is Chromium might
fix the issue.