fix: Fix flameshot clipboard copy-image issue
I would have a weird issue where I would sometimes fail to use Flameshot
to take a screenshot of a part of the screen and paste it into Anki. It
appears it would only happen if I selected a large enough area of the
screen.

It looks like the issue was caused by the fact that
`QT_QPA_PLATFORM=wayland` was not set at the moment Sway is launched. I
had previously avoided exporting this envvar since it would cause Anki
to crash. But now that I have fixed the Anki issue, I can set it and get
Flameshot to work too.