fix: Adjust toggle_handsfree.sh to Pipewire
Pipewire (or Wireplumber?) will automatically switch between HFP mode
and A2DP mode for my Bluetooth headset depending on if I am using an
application that requests using a mic, such as a mic testing website.
This doesn't work perfectly however. I can't get it to work in Teams
sometimes. So I rely on this script as a backup.
With Pipewire, the system is switching between these two profiles:

headset-head-unit-msbc
a2dp-sink

However, additional profiles are shown in the output of `pactl list
cards`. So I needed to blacklist those, or else the script wouldn't
work with my new Pipewire setup.