fix: Fix issue where mousemapper daemon takes time to start after boot
There is a bug where, after rebooting my system, attempting to switch
tabs in Chromium using my MX Master mouse would not work. I would hold
the Smart Shift button while using the thumb wheel to scroll left. This
would initially fail, but start working again after 5 seconds.

It turns out that the mousemapper daemon is started a bit too early,
before the rest of the system is properly initialized (in particular,
the `swaymsg` is unusable in that state). That, in turn, caused the
mousemapper daemon to crash, wait 5 seconds (due to `RestartSec=5s`),
then restart, and then work again.

To work around this problem, I am attempting to use the `SWAYSOCK`
environment variable early in the script. If the daemon was started too
early, this variable will be unset, triggering an early exit of the
daemon (and subsequent restart) due to `set -u`. When the daemon
restarts, it will now have `SWAYSOCK` set, and the `swaymsg` will work
correctly.