feat: Invert Smart Shift behavior
How often do I feel the need to switch tabs in Chromium? All the time.
How often do I feel the need to rewind a youtube video. More seldom.
What is easier: to hold Smart Shift while scrolling the thumbwheel, or
to just scroll the thumbwheel? To just scroll the thumbwheel. It
follows, then, that switching tabs should be the easier operation.

The main drawback is that it will become a bit more tricky to scroll
horizontally in e.g. Visio drawings in the browser. But I'm not doing
that so often. And I'm using the drawio desktop app for my diagrams, not
Chromium. If I find myself in such a situation, I could just hold the
Smart Shift button while scrolling. Or I could adjust the script so that
it will refuse to switch tabs if the window name contains the word
"Visio" or something.